+++ CHANGELOG.md
@@ -4,6 +4,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [0.5.0] - 2026-07-27
+
+- doc(README): updated installed instructions
+- change: bin/nt_powerdns, use @dns-rr
+- change: bin/nt_powerdns: moved to @nictool/dns-nameserver
+- change: default file format is JSON (0 dep)
+- change: NS config removed to API
+
 ### [0.2.2] - 2026-07-25
 
 - test and start fixes